feat(kit): add TuiProgressColorSegmentsAsyncPipe
