+++ projects/kit/components/progress/index.ts
@@ -3,4 +3,6 @@ export * from './progress-bar/progress-bar.component';
 export * from './progress-circle/progress-circle.component';
 export * from './progress-segmented/progress-segmented.component';
 export * from './progress-label/progress-label.component';
+export * from './pipes/abstract-color-segments';
 export * from './pipes/progress-color-segments.pipe';
+export * from './pipes/progress-color-segments-async.pipe';
