+++ projects/kit/components/progress/pipes/progress-color-segments.pipe.ts
@@ -1,33 +1,14 @@
-import {ElementRef, Inject, Pipe, PipeTransform} from '@angular/core';
-import {USER_AGENT} from '@ng-web-apis/common';
-import {CHROMIUM_EDGE_START_VERSION, isEdgeOlderThan, isIE} from '@taiga-ui/cdk';
+import {Pipe} from '@angular/core';
+import {AbstractTuiColorSegments} from './abstract-color-segments';
 
-@Pipe({
-    name: 'tuiProgressColorSegments',
-})
-export class TuiProgressColorSegmentsPipe implements PipeTransform {
-    private readonly isOldBrowsers =
-        isEdgeOlderThan(CHROMIUM_EDGE_START_VERSION, this.userAgent) ||
-        isIE(this.userAgent);
-
-    constructor(
-        @Inject(USER_AGENT) private readonly userAgent: string,
-        @Inject(ElementRef) private readonly elementRef: ElementRef<HTMLElement>,
-    ) {}
-
+/**
+ * @deprecated use tuiProgressColorSegmentsAsync instead
+ * {@link TuiProgressColorSegmentsAsyncPipe}
+ * example: `<progress [color]="colors | tuiProgressColorSegmentsAsync | async">`
+ */
+@Pipe({name: 'tuiProgressColorSegments'})
+export class TuiProgressColorSegmentsPipe extends AbstractTuiColorSegments<string> {
     transform(colors: string[]): string {
-        if (this.isOldBrowsers) {
-            return colors[0];
-        }
-
-        const elementWidth = this.elementRef.nativeElement.offsetWidth;
-        const segmentWidth = Math.ceil(elementWidth / colors.length);
-        const colorsString = colors.reduce(
-            (acc, color, i) =>
-                `${acc}, ${color} ${i * segmentWidth}px ${(i + 1) * segmentWidth}px`,
-            '',
-        );
-
-        return `linear-gradient(to right ${colorsString})`;
+        return this.calculate(colors);
     }
 }
